+++ src/PatientModel.java
@@ -1,16 +1,142 @@
+import java.util.Arrays;
 
 /**handles all calculations, database queries, and the overall work needed to be done for handling this associated role
  * Does NOT interact with the view class directly, and also does NOT interact with the Controller class(The Controller
  * class interacts with this class, not the other way around.)*/
 public class PatientModel extends UserModel {
 
+	private String address;
+	private String phoneNumber;
+	private String email;
+	private char[] password2;
+	private float amountDue;
+	
 
 	//constructor
-	public PatientModel(String username, char[] password, String name) {
+	public PatientModel(String username, char[] password, String name
+			, String addr, String phNumber, String email, float amountDue)
+	{
 		
 		super(name,username,password);
 		setRole("patient");
+		this.address = addr;
+		this.phoneNumber = phNumber;
+		this.email = email;
+		this.amountDue = amountDue;
+	}
+	
+	
+	
+	
+	
+	
+	//convert a string of digits to a phone number
+	public String convertToPhoneNumber()
+	{
+		//https://howtodoinjava.com/java/string/format-phone-number/
+		return  phoneNumber.replaceFirst("(\\d{3})(\\d{3})(\\d+)", "($1) $2-$3");
+	}
+	
+	//convert a float to a String with 2 decimal places
+	public String convertToDollar()
+	{
+		return String.format("$%.2f" , amountDue);
+	}
+
+	public String updatePassword()
+	{
+		//check that password is long enough; pwd's already verified at this point
+		if(password.length < 4)
+		{
+			return "Password must be at least 4 characters!";
+		}
 		
+		//check if passwords match
+		if(Arrays.equals(password, password2) == false)
+		{
+			return "Passwords don't match!";
+		}
+		
+		//else tests pass, store password for the user
+		Main.dbase.get(username).setPassword(password);
+		
+		return "Password Changed Successfully!";
+		
+		
+	}
+	
+	
+//TODO error checking on instance variables set by model.updateInfo()
+	public String verifyInfo()
+	{
+		//error checks, don't update database in here
+		if(true)
+		{
+			
+		}
+		
+		//update the information in the dbase once all checks pass
+		return "Information Successfully Updated";
+	}
+	
+	
+	
+/**Getters and Setters*/
+	public String getAddress() {
+		return address;
+	}
+
+
+	public void setAddress(String address) {
+		this.address = address;
+	}
+
+
+	public String getPhoneNumber() {
+		return phoneNumber;
+	}
+
+
+	public void setPhoneNumber(String phoneNumber) {
+		this.phoneNumber = phoneNumber;
+	}
+
+
+	public String getEmail() {
+		return email;
+	}
+
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
+
+	public float getAmountDue() {
+		return amountDue;
+	}
+
+
+	public void setAmountDue(float amountDue) {
+		this.amountDue = amountDue;
+	}
+
+
+
+
+
+
+	public char[] getPassword2() {
+		return password2;
+	}
+
+
+
+
+
+
+	public void setPassword2(char[] password2) {
+		this.password2 = password2;
 	}