+++ CHANGELOG.md
@@ -3,7 +3,26 @@ __In descending date order first__ (most recent date and time first).
 Timestamp Format: __day month HOURS:MINUTES (military time) USER__  
 Example: 10 mar 21:13 SAJID
 
-##15 mar 00:30 Sajid
+
+
+## 16 mar 06:01 Sajid
+**patient portal functionality created**
+* uploads patient information from the database
+* checks that no fields are empty when save button is pressed
+* doesnt do error checking or update the database on 'patient info' box yet when save button is pressed-- maybe leave it for next iteration
+* patient credentials panel does error checking and updates the database to the changed password
+**databaseModel updated**
+* new fields added" address, ph#, email, amountDue
+**dbase.txt updated with new fields**
+* new fields added" address, ph#, email, amountDue
+**PatientModel**
+* constructor changed to reflect changes above
+**CreateNewPatientMVC**
+* refactored View using MigLayout
+* Added fields address, email, phone number which store in database as is
+* changed the controller and model to reflect these new fields
+
+## 15 mar 00:30 Sajid
 * added 'create department' functionality for admin
 
 ## 15 mar 15:50 Sajid